refactor(lint): 完善lint
