--- src/Button/index.tsx
@@ -1,4 +1,0 @@
-import Button from './Button';
-import Group from './Group';
-
-Button.Group = Group;
