+++ tsconfig.json
@@ -9,6 +9,7 @@
     "sourceMap": true,
     "baseUrl": "./",
     "strict": true,
+    "removeComments": true,
     "paths": {
       "@/*": ["src/*"],
       "@@/*": ["src/.umi/*"]
